Backend/requirements.txt
fastapi
uvicorn
pydantic
motor
langchain
langchain-groq
langchain-community
langgraph
tavily-python
python-dotenv
mangum